refactor(hooks): path
