+++ packages/shared/hooks/useConfig/index.ts
@@ -1,6 +1,6 @@
 import { computed } from 'vue';
 // TODO need refactor
-import { useConfig } from '../../components/config-provider/hooks/useConfig';
+import { useConfig } from '../../../components/config-provider/hooks/useConfig';
 
 export function usePrefixClass(componentName?: string) {
   const { classPrefix } = useConfig('classPrefix');
