+++ packages/shared/hooks/useGlobalIcon/index.ts
@@ -1,6 +1,6 @@
 import { useConfig } from '../useConfig';
 // TODO need refactor
-import { IconConfig } from '../../components/config-provider/type';
+import { IconConfig } from '../../../components/config-provider/type';
 
 // 从 globalConfig 获取 icon 配置用于覆盖组件内置 icon
 export function useGlobalIcon(tdIcon: object) {
