+++ packages/shared/hooks/useReadonly/index.ts
@@ -1,7 +1,7 @@
 import { Ref, inject, computed, getCurrentInstance } from 'vue';
 import { isBoolean } from 'lodash-es';
 // TODO need refactor
-import { TdFormProps } from '../../components/form/type';
+import { TdFormProps } from '../../../components/form/type';
 
 export interface FormReadonlyProvider {
   readonly: Ref<TdFormProps['readonly']>;
